training/trainer.py
"""
Training Loop

The main training orchestrator that ties together:
    - Data loading
    - Model forward/backward
    - Optimizer and scheduler
    - Mixed precision
    - Gradient accumulation
    - Gradient checkpointing
    - Logging and checkpointing
    - Evaluation

This is the central training script — no Trainer class imported
from any framework. Everything is explicit.

Usage:
    trainer = Trainer(config)
    trainer.train()
"""

import os
import time
import yaml
import torch
from typing import Optional

from model.config import ModelConfig
from model.transformer import Transformer
from data.loader import DataLoader
from training.adamw import AdamW
from training.muon import Muon
from training.scheduler import build_scheduler
from training.precision import FP32MasterWeights, GradScaler
from training.grad_accumulation import GradientAccumulator
from training.loss import compute_total_loss
from utils.logging import TrainingLogger
from utils.checkpoint import save_checkpoint, load_checkpoint


class Trainer:
    """
    Training orchestrator.

    Manages the full training pipeline from config to trained model.

    Args:
        config_path: Path to YAML configuration file.
        resume_from: Optional checkpoint path to resume from.
    """

    def __init__(self, config_path: str, resume_from: Optional[str] = None):
        # Load config
        with open(config_path, "r") as f:
            self.raw_config = yaml.safe_load(f)

        self.model_config = ModelConfig.from_dict(self.raw_config["model"])
        self.train_config = self.raw_config["training"]
        self.optim_config = self.raw_config["optimizer"]
        self.log_config = self.raw_config.get("logging", {})
        self.ckpt_config = self.raw_config.get("checkpointing", {})

        # Device
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

        # Initialize components (done in setup)
        self.model: Optional[Transformer] = None
        self.optimizer = None
        self.scheduler = None
        self.train_loader: Optional[DataLoader] = None
        self.val_loader: Optional[DataLoader] = None
        self.logger: Optional[TrainingLogger] = None
        self.grad_accumulator: Optional[GradientAccumulator] = None
        self.grad_scaler: Optional[GradScaler] = None

        self.global_step = 0
        self.tokens_seen = 0
        self.resume_from = resume_from

    def setup(self) -> None:
        """Initialize all training components."""
        # TODO: Implement full setup
        # 1. Build model from config
        # 2. Move to device
        # 3. Print parameter count
        # 4. Build optimizer (AdamW or Muon based on config)
        # 5. Build scheduler
        # 6. Build data loaders
        # 7. Setup mixed precision (FP32 master weights, grad scaler)
        # 8. Setup gradient accumulation
        # 9. Setup logger
        # 10. Optionally resume from checkpoint
        raise NotImplementedError

    def train(self) -> None:
        """
        Main training loop.

        For each step:
            1. Get batch from data loader
            2. Forward pass (with mixed precision)
            3. Compute loss
            4. Backward pass (scaled gradients)
            5. Gradient accumulation check
            6. If accumulation complete: clip gradients, optimizer step, scheduler step
            7. Log metrics
            8. Periodic evaluation
            9. Periodic checkpointing
        """
        # TODO: Implement main training loop
        raise NotImplementedError

    def _train_step(self, batch: dict) -> dict:
        """
        Single training micro-step.

        Returns dict with loss values and metrics.
        """
        # TODO: Implement single training step
        raise NotImplementedError

    @torch.no_grad()
    def evaluate(self) -> dict:
        """
        Run evaluation on validation set.

        Returns dict with eval metrics (loss, perplexity, etc.).
        """
        # TODO: Implement evaluation loop
        raise NotImplementedError

    def save(self, tag: str = "latest") -> None:
        """Save training checkpoint."""
        # TODO: Implement checkpoint saving
        raise NotImplementedError

    def load(self, path: str) -> None:
        """Load training checkpoint and resume."""
        # TODO: Implement checkpoint loading
        raise NotImplementedError

    def _log_step(self, step: int, metrics: dict, lr: float, elapsed: float) -> None:
        """Log training metrics for one step."""
        # TODO: Implement step logging
        raise NotImplementedError

    def _print_config(self) -> None:
        """Print training configuration summary."""
        # TODO: Implement config printing
        raise NotImplementedError
