pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "swahili-100"
version = "0.1.0"
description = "A from-scratch Swahili language model under $100"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Marco Haruni"},
]
keywords = ["nlp", "swahili", "language-model", "transformer", "from-scratch"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.2.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
logging = [
    "wandb>=0.16.0",
]
export = [
    "onnx>=1.15.0",
    "onnxruntime>=1.17.0",
]
gpu = [
    "triton>=2.1.0",
]

[project.urls]
Homepage = "https://github.com/marcoharuni/swahili-100"
Repository = "https://github.com/marcoharuni/swahili-100"
Issues = "https://github.com/marcoharuni/swahili-100/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.setuptools.packages.find]
include = ["model*", "data*", "training*", "evaluation*", "alignment*", "quantization*", "utils*"]
